# cmd/shortener/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сокращения URL.",
        "title": "URLShortener API",
        "contact": {
            "email": "arsenzhar@yandex.ru."
        },
        "version": "1.0."
    },
    "host": "localhost:8080.",
    "basePath": "/",
    "paths": {
...
        "/api/shorten/batch": {
            "post": {
                "description": "Получает на вход массив ссылок и отдаёт в ответе сокращённый вариант",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Сокращает массив ссылок",
                "parameters": [
                    {
                        "description": "Массив URL для сокращения",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RequestLinks"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseLinks"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RequestLinks": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "CorrelationID – ID ссылки.",
                    "type": "string",
                    "format": "uuid",
                    "example": "58039b0a-480d-11ee-9ace-0e6250a0eb02"
                },
                "original_url": {
                    "description": "OriginalURL – ссылка, которую необходимо сократить.",
                    "type": "string",
                    "example": "https://ya.ru"
                }
            }
        },
...
        "models.ResponseLinks": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "CorrelationID – ID ссылки.",
                    "type": "string",
                    "format": "uuid",
                    "example": "58039b0a-480d-11ee-9ace-0e6250a0eb02"
                },
                "short_url": {
                    "description": "ShortURL – сокращённая ссылка.",
                    "type": "string",
                    "example": "http://localhost:8080/maIJa1"
                }
            }
        }
    }
}